fix : inventory stacking problem, dynamic sprite change for oreNode when it's mined
- add sprite sheet constants for ore visualization
- configure resource node richness scaling based on world parameters
- inject World dependency into ResourceNodeSystem for ore amount access
- optimize resource mining logic to prevent over-collection

+++ include/Components/ResourceNodeComponent.h
@@ -3,7 +3,7 @@
 
 #include "Core/Item.h"
 
-using rsrc_amt_t = unsigned long long;
+using rsrc_amt_t = unsigned int;
 
 struct ResourceNodeComponent {
   rsrc_amt_t LeftResource;
